refactor: move coordinator helpers into package
Consolidate all root-level _coordinator_*.py and coordinator_*.py helper
modules into custom_components/thessla_green_modbus/coordinator/ making it
the single canonical home for coordinator logic.

Files moved (17 total):
  _coordinator_connection_test.py  → coordinator/connection_test.py
  _coordinator_device_info.py      → coordinator/device_info.py
  _coordinator_factory.py          → coordinator/factory.py
  _coordinator_init.py             → coordinator/init_config.py
  _coordinator_read_batches.py     → coordinator/read_batches.py
  _coordinator_read_bits.py        → coordinator/read_bits.py
  _coordinator_read_common.py      → coordinator/read_common.py
  _coordinator_register_groups.py  → coordinator/register_groups.py
  _coordinator_register_processing.py → coordinator/register_processing.py
  _coordinator_runtime_io.py       → coordinator/runtime_io.py
  _coordinator_scan_result.py      → coordinator/scan_result.py
  _coordinator_scanner_kwargs.py   → coordinator/scanner_kwargs.py
  _coordinator_transport_select.py → coordinator/transport_select.py
  coordinator_config.py            → coordinator/config_normalization.py
  coordinator_diagnostics.py       → coordinator/diagnostics.py
  coordinator_runtime.py           → coordinator/runtime.py
  coordinator_state.py             → coordinator/state.py

Import updates:
  coordinator/coordinator.py: all .._ and ..coordinator_ imports → intra-package
  coordinator/io.py: all .._ imports → intra-package
  coordinator/models.py: lazy ..coordinator_config → .config_normalization
  Moved files: .const/.modbus_exceptions/etc. → ..const/..modbus_exceptions/etc.
               .coordinator.retry → .retry, .coordinator.models → .models
  Tests: test_coordinator_availability, test_coordinator_device_info,
         test_coordinator_scan_cache_extra, test_coordinator_schedule,
         test_optimized_integration_updates

No compatibility shims needed; all callers updated in this PR.
Behavior, entity IDs, service IDs, and register names are unchanged.
