+++ custom_components/thessla_green_modbus/coordinator/connection_test.py
@@ -5,8 +5,8 @@ from __future__ import annotations
 from collections.abc import Awaitable, Callable, Iterable
 from typing import Any
 
-from .modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
-from .transport.base import BaseModbusTransport
+from ..modbus_exceptions import ConnectionException, ModbusException, ModbusIOException
+from ..transport.base import BaseModbusTransport
 
 
 async def run_connection_test(
