+++ custom_components/thessla_green_modbus/coordinator/init_config.py
@@ -6,8 +6,8 @@ from collections.abc import Callable
 from datetime import timedelta
 from typing import Any
 
-from .const import CONNECTION_MODE_AUTO
-from .coordinator.models import CoordinatorConfig
+from ..const import CONNECTION_MODE_AUTO
+from .models import CoordinatorConfig
 
 
 def normalize_runtime_config(
