+++ custom_components/thessla_green_modbus/coordinator/read_batches.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from .coordinator.retry import _PermanentModbusError
-from .modbus_exceptions import ConnectionException, ModbusException
-from .registers.read_planner import chunk_register_range
+from ..modbus_exceptions import ConnectionException, ModbusException
+from ..registers.read_planner import chunk_register_range
+from .retry import _PermanentModbusError
 
 _LOGGER = logging.getLogger(__name__)
