+++ custom_components/thessla_green_modbus/coordinator/register_groups.py
