+++ custom_components/thessla_green_modbus/coordinator/runtime_io.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 from typing import Any, cast
 
-from .modbus.call import _call_modbus
-from .modbus_exceptions import ConnectionException
+from ..modbus.call import _call_modbus
+from ..modbus_exceptions import ConnectionException
 
 
 async def call_modbus(
