+++ custom_components/thessla_green_modbus/coordinator/scan_result.py
