+++ custom_components/thessla_green_modbus/coordinator/scanner_kwargs.py
